src/requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
astroid==2.0.4
autopep8==1.4
billiard==3.5.0.4
boto==2.49.0
boto3==1.8.3
botocore==1.11.3
celery==4.2.1
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
diff-match-patch==20121119
Django==2.1.2
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-import-export==1.0.1
django-storages==1.6.6
django-templated-mail==1.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
et-xmlfile==1.0.1
geographiclib==1.49
geopy==1.17.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jmespath==0.9.3
kombu==4.2.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.15.1
odfpy==1.3.6
openpyxl==2.5.5
pandas==0.23.4
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
tablib==0.12.1
typed-ast==1.1.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
